export result to separate file
